+++ common/src/main/java/hungteen/craid/CRaidPlatformAPI.java
@@ -3,8 +3,14 @@ package hungteen.craid;
 import hungteen.craid.api.capability.RaidCapability;
 import hungteen.craid.common.world.raid.HTRaidImpl;
 import hungteen.htlib.api.util.helper.ServiceHelper;
+import net.minecraft.world.DifficultyInstance;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.MobSpawnType;
+import net.minecraft.world.entity.SpawnGroupData;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.ServerLevelAccessor;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.Optional;
 
@@ -37,5 +43,9 @@ public interface CRaidPlatformAPI {
     /* Cap Related */
 
     Optional<? extends RaidCapability> getRaidCap(Entity entity);
+
+    /* Misc */
+
+    SpawnGroupData onFinalizeSpawn(Mob mob, ServerLevelAccessor level, DifficultyInstance difficulty, MobSpawnType spawnType, @Nullable SpawnGroupData spawnData);
 
 }