+++ fabric/src/main/java/hungteen/craid/CRaidFabricAPI.java
@@ -4,8 +4,14 @@ import hungteen.craid.api.capability.RaidCapability;
 import hungteen.craid.api.event.RaidEvent;
 import hungteen.craid.common.CRaidEntityComponents;
 import hungteen.craid.common.world.raid.HTRaidImpl;
+import net.minecraft.world.DifficultyInstance;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.MobSpawnType;
+import net.minecraft.world.entity.SpawnGroupData;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.ServerLevelAccessor;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.Optional;
 
@@ -34,5 +40,10 @@ public class CRaidFabricAPI implements CRaidPlatformAPI {
     @Override
     public Optional<? extends RaidCapability> getRaidCap(Entity entity) {
         return Optional.of(CRaidEntityComponents.RAID.get(entity));
+    }
+
+    @Override
+    public SpawnGroupData onFinalizeSpawn(Mob mob, ServerLevelAccessor level, DifficultyInstance difficulty, MobSpawnType spawnType, @Nullable SpawnGroupData spawnData) {
+        return mob.finalizeSpawn(level, difficulty, spawnType, spawnData);
     }
 }