+++ neoforge/src/main/java/hungteen/craid/common/attachment/CRaidAttachments.java
@@ -1,17 +1,13 @@
 package hungteen.craid.common.attachment;
 
 import hungteen.craid.api.CRaidAPI;
+import hungteen.craid.common.capability.RaidAttachmentSerializer;
 import hungteen.craid.common.capability.RaidCapabilityImpl;
-import net.minecraft.core.HolderLookup;
-import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.entity.Entity;
 import net.neoforged.neoforge.attachment.AttachmentType;
-import net.neoforged.neoforge.attachment.IAttachmentHolder;
-import net.neoforged.neoforge.attachment.IAttachmentSerializer;
 import net.neoforged.neoforge.registries.DeferredHolder;
 import net.neoforged.neoforge.registries.DeferredRegister;
 import net.neoforged.neoforge.registries.NeoForgeRegistries;
-import org.jetbrains.annotations.Nullable;
 
 /**
  * @program: CustomRaid
@@ -27,23 +23,10 @@ public interface CRaidAttachments {
             RaidCapabilityImpl capability = new RaidCapabilityImpl();
             if(holder instanceof Entity entity){
                 capability.init(entity);
+            } else {
+                throw new IllegalArgumentException("Raid capability can only be attached to Entity");
             }
             return capability;
-        }).serialize(new IAttachmentSerializer<CompoundTag, RaidCapabilityImpl>() {
-            @Override
-            public RaidCapabilityImpl read(IAttachmentHolder holder, CompoundTag tag, HolderLookup.Provider provider) {
-                RaidCapabilityImpl capability = new RaidCapabilityImpl();
-                if(holder instanceof Entity entity){
-                    capability.init(entity);
-                }
-                capability.deserializeNBT(provider, tag);
-                return capability;
-            }
-
-            @Override
-            public @Nullable CompoundTag write(RaidCapabilityImpl capability, HolderLookup.Provider provider) {
-                return capability.serializeNBT(provider);
-            }
-        }).build();
+        }).serialize(new RaidAttachmentSerializer()).build();
     });
 }