+++ docs/en_us/changelog.md
@@ -1,4 +1,4 @@
 ## 1.21.x
 ### 1.21.1-1.0.0
 #### New Features
-* 
+* Support multiple platforms, providing support for forge, neoforge, and fabric.