+++ docs/zh_cn/changelog.md
@@ -1,4 +1,4 @@
 ## 1.21.x
 ### 1.21.1-1.0.0
 #### 新内容
-* 多平台支持，提供对 forge、neoforge 的支持。
+* 多平台支持，提供对 forge、neoforge、fabric 的支持。